feat(Bonus Pagamenti Digitali): [#176071117] Refactor Bancomat analytics (#2572)
* [#176071117] create the bancomat analytics file

* [#176071117] add bancomat analytics to the middleware

* [#176071117] remove bancomat action from the bpd analytics

* [#176071117] remove unused import
